+++ lib/types/index.ts
@@ -7,6 +7,8 @@ export * from './CreateClientOptions';
 export * from './CreateOptions';
 export * from './FailedItem';
 export * from './ObjectValue';
+export * from './PathLeaf';
+export * from './PathNode';
 export * from './RequestCallback';
 export * from './RequestOptions';
 export * from './SmartsheetClient';